requirements.txt
# FastAPI + server
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0

# Vector DB + cache
qdrant-client>=1.13.0
redis>=5.2.0

# Embedding + reranking
voyageai>=0.3.0
cohere>=5.13.0

# Document conversion
pymupdf4llm>=0.0.17
pymupdf>=1.25.0
python-docx>=1.1.0
openpyxl>=3.1.0
python-pptx>=1.0.0
pytesseract>=0.3.13
Pillow>=11.0.0
imagehash>=4.3.0

# LLM / Vision APIs
google-genai>=1.0.0
openai>=1.60.0
anthropic>=0.28.0

# NLP utilities
langdetect>=1.0.9
tiktoken>=0.8.0

# Config
pyyaml>=6.0.2

# Misc
httpx>=0.28.0
python-multipart>=0.0.18
python-magic>=0.4.27
uuid6>=2024.7.10
rich>=13.7.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
